+++ WalletWasabi.Fluent/Models/Wallets/WalletModel.cs
@@ -137,35 +137,3 @@ public partial class WalletModel : ReactiveObject
 			.Select(x => new Address(Wallet.KeyManager, x));
 	}
 }
-
-[AutoInterface]
-public partial class WalletCoinsModel
-{
-	private readonly Wallet _wallet;
-
-	public WalletCoinsModel(Wallet wallet, IWalletModel walletModel)
-	{
-		_wallet = wallet;
-
-		List =
-			Observable.Defer(() => GetCoins().ToObservable())                                                        // initial coin list
-					  .Concat(walletModel.TransactionProcessed.SelectMany(_ => GetCoins()))                          // Refresh whenever there's a relevant transaction
-					  .Concat(walletModel.WhenAnyValue(x => x.Settings.AnonScoreTarget).SelectMany(_ => GetCoins())) // Also refresh whenever AnonScoreTarget changes
-					  .ToObservableChangeSet();
-
-		Pockets =
-			Observable.Defer(() => _wallet.GetPockets().ToObservable())                                                       // initial pocket list
-					  .Concat(walletModel.TransactionProcessed.SelectMany(_ => wallet.GetPockets()))                          // Refresh whenever there's a relevant transaction
-					  .Concat(walletModel.WhenAnyValue(x => x.Settings.AnonScoreTarget).SelectMany(_ => wallet.GetPockets())) // Also refresh whenever AnonScoreTarget changes
-					  .ToObservableChangeSet();
-	}
-
-	public IObservable<IChangeSet<ICoinModel>> List { get; }
-
-	public IObservable<IChangeSet<Pocket>> Pockets { get; }
-
-	private IEnumerable<ICoinModel> GetCoins()
-	{
-		return _wallet.Coins.Select(x => new CoinModel(_wallet, x));
-	}
-}
